Teach tctl to use remote auth servers and identity.
